requirements.txt
pandas
pytest
streamlit
numpy
python-dotenv
openai
fastapi
uvicorn
requests
--extra-index-url https://pypi.fury.io/ericmichael/
omniagents[all]>=0.6.31
pyyaml
